requirements.txt
crewai[anthropic]==1.9.3
crewai-tools==1.9.3
python-dotenv>=1.0.0
